cleans code

+++ model/option.go
@@ -6,7 +6,7 @@ import "time"
 type Option struct {
 	ID         uint64   `gorm:"AUTO_INCREMENT;primary_key"`
 	Question   Question `gorm:"foreignkey:QuestionID"`
-	QuestionID uint
+	QuestionID uint64
 	OptText    string    `gorm:"type:varchar(255);not null"`
 	CreatedAt  time.Time `gorm:"default:'now()';not null"`
 	UpdatedAt  time.Time `gorm:"default:'now()'"`
